Minor fixes for .vox support.

+++ llc/gpk_view_array.h
@@ -161,9 +161,9 @@ namespace gpk
 
 	template<typename _tElement>
 						::gpk::error_t			find								(const _tElement& element, const ::gpk::view_array<const _tElement>& target, uint32_t offset = 0)				{
-		for(int32_t iOffset = offset, offsetStop = target.size(); iOffset < offsetStop; ++iOffset)
+		for(uint32_t iOffset = offset, offsetStop = target.size(); iOffset < offsetStop; ++iOffset)
 			if(element == target[iOffset])
-				return iOffset;
+				return (int32_t)iOffset;
 		return -1;
 	}
